+++ src/Greenleaf/ActionTrait.php
@@ -10,11 +10,11 @@ declare(strict_types=1);
 namespace DecodeLabs\Greenleaf;
 
 use DecodeLabs\Harvest\Request as HarvestRequest;
+use DecodeLabs\Pandora\Container as PandoraContainer;
 use DecodeLabs\Singularity\Url\Leaf as LeafUrl;
+use DecodeLabs\Slingshot;
+use Psr\Container\ContainerInterface as Container;
 use Psr\Http\Message\ServerRequestInterface as Request;
-use Psr\Http\Message\UriInterface as Uri;
-use ReflectionClass;
-use ReflectionNamedType;
 
 trait ActionTrait
 {
@@ -31,84 +31,53 @@ trait ActionTrait
 
 
     /**
-     * Prepare method parameters
-     *
-     * @return array<string, mixed>
+     * Get middleware list
      */
-    protected function prepareMethodParameters(
-        string $method,
+    public function getMiddleware(): ?array
+    {
+        if (!defined('static::MIDDLEWARE')) {
+            return null;
+        }
+
+        return static::MIDDLEWARE;
+    }
+
+
+
+    /**
+     * Prepare slingshot
+     */
+    protected function prepareSlingshot(
         array $parameters,
         LeafUrl $url,
         Request $request
-    ): array {
-        $ref = new ReflectionClass($this);
-        $method = $ref->getMethod($method);
-        $params = $method->getParameters();
-        $query = $url->parseQuery();
-        $queryParams = $request->getQueryParams();
-        $output = [];
+    ): Slingshot {
+        $output = new Slingshot(
+            container: $this->context->container
+        );
 
-        foreach ($params as $param) {
-            $name = $param->getName();
+        $output->addParameters($request->getAttributes());
+        $output->addParameters($request->getQueryParams());
+        $output->addParameters($url->parseQuery()->toArray());
+        $output->addParameters($parameters);
 
-            // Parameters passed in from route
-            if (isset($parameters[$name])) {
-                $output[$name] = $parameters[$name];
-                continue;
-            }
+        $output->addTypes([
+            LeafUrl::class => $url,
+            Request::class => $request,
+            Container::class => $this->context->container
+        ]);
 
-            // Leaf URL query
-            if (isset($query[$name])) {
-                $output[$name] = $query[$name];
-                continue;
-            }
+        if ($request instanceof HarvestRequest) {
+            $output->addType(
+                $request,
+                HarvestRequest::class
+            );
+        }
 
-            // PSR-7 query
-            if (array_key_exists($name, $queryParams)) {
-                $output[$name] = $queryParams[$name];
-                continue;
-            }
-
-            // Type
-            if (
-                ($type = $param->getType()) &&
-                $type instanceof ReflectionNamedType
-            ) {
-                switch ($type) {
-                    case HarvestRequest::class:
-                        if (!$request instanceof HarvestRequest) {
-                            break;
-                        }
-
-                        // no break
-                    case Request::class:
-                        $output[$name] = $request;
-                        continue 2;
-
-                    case LeafUrl::class:
-                        $output[$name] = $url;
-                        continue 2;
-
-                    case Uri::class:
-                        $output[$name] = $request->getUri();
-                        continue 2;
-                }
-            }
-
-            // Default value
-            if ($param->isDefaultValueAvailable()) {
-                $output[$name] = $param->getDefaultValue();
-                continue;
-            }
-
-            // Optional parameter
-            if ($param->isOptional()) {
-                $output[$name] = null;
-                continue;
-            }
-
-            throw Exceptional::UnexpectedValue(
-                'Method "' . $method->getName() . '" missing required parameter: ' . $name
+        if ($this->context->container instanceof PandoraContainer) {
+            $output->addType(
+                $this->context->container,
+                PandoraContainer::class
             );
         }