+++ CHANGELOG.md
@@ -1,2 +1,8 @@
+## v0.1.1 (2023-11-27)
+* Added Middleware support to Actions
+* Use request attributes in Action Slingshot
+* Switched to Slingshot for invocation
+* Updated Dictum dependency
+
 ## v0.1.0 (2023-11-09)
 * Built initial implementation